third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-offset-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'offset' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(feRFunc.offset.baseVal * 100) is 50
PASS Math.round(feGFunc.offset.baseVal * 100) is 50
PASS Math.round(feBFunc.offset.baseVal * 100) is 50
PASS Math.round(feAFunc.offset.baseVal * 100) is 50
PASS Math.round(feRFunc.offset.baseVal * 100) is 10
PASS Math.round(feGFunc.offset.baseVal * 100) is 10
PASS Math.round(feBFunc.offset.baseVal * 100) is 10
PASS Math.round(feAFunc.offset.baseVal * 100) is 10
PASS successfullyParsed is true

TEST COMPLETE
